[MerkleTreeApi] Add Models/ for Request/Response Schema
- Moved Create/GetUserBalance Request/Response schema into that folder
