+++ MerkleTreeApi/Models/CreateResponse.cs
@@ -1,0 +1,5 @@
+namespace MerkleTreeApi.Models;
+
+public class CreateResponse {
+    public string root_hash { get; set; } = "";
+}
