+++ MerkleTreeApi/Models/GetBalanceRequest.cs
@@ -1,0 +1,5 @@
+namespace MerkleTreeApi.Models;
+
+public class GetBalanceRequest {
+    public int user_id { get; set; }
+}
