+++ MerkleTreeApi/Models/GetBalanceResponse.cs
@@ -1,0 +1,6 @@
+namespace MerkleTreeApi.Models;
+
+public class GetBalanceResponse {
+    public int balance { get; set; }
+    public MerkleTree.MerkleProofNode[] proofs { get; set; }
+}
